refracting changes
