+++ backend/src/main/java/com/smartroom/backend/controller/TeacherController.java
@@ -6,12 +6,16 @@ import com.smartroom.backend.entity.StudentDetails;
 import com.smartroom.backend.model.StudentModel;
 import com.smartroom.backend.service.TeacherService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.mongodb.core.MongoOperations;
+import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -24,6 +28,8 @@ public class TeacherController {
 
     private final EmailSender emailSender;
 
+
+
     @Autowired
     TeacherController(TeacherService teacherService,EmailSender emailSender) {
 
@@ -34,6 +40,7 @@ public class TeacherController {
 
     @PostMapping("/create/student")
     public ResponseEntity<StudentModel> createStudent(@RequestBody @Valid Student student) throws Exception {
+        System.out.println("student "+student);
         StudentModel createdStudent = teacherService.createStudent(student);
         return new ResponseEntity<>(createdStudent, HttpStatus.CREATED);
     }
@@ -45,11 +52,6 @@ public class TeacherController {
         return new ResponseEntity<>(updatedStudent, HttpStatus.CREATED);
     }
 
-    @GetMapping("/student/fetch/all")
-    public ResponseEntity<List<Student>> fetchAllStudent() throws Exception {
-        List<Student> fetchedStudents = teacherService.fetchAllStudent();
-        return new ResponseEntity<>(fetchedStudents, HttpStatus.OK);
-    }
 
     @GetMapping("/predict/{studentId}/{subject}")
     public ResponseEntity<String> predictStudentResult(@PathVariable("studentId") String studentId, @PathVariable("subject") String subject) throws Exception {
@@ -60,9 +62,11 @@ public class TeacherController {
 
     @PostMapping("send/email/{studentMail}")
     public ResponseEntity<String> sendEmail(@RequestBody String message ,@PathVariable("studentMail") String studentMail){
-         emailSender.sendEmail(studentMail,message,"Attention!! Predicted Score");
+         emailSender.sendEmail(studentMail,message,"Attention Needed!!");
          return new ResponseEntity<>("Mail Successfully Send" , HttpStatus.OK);
     }
 
+    }
 
-}
+
+
