+++ backend/src/main/java/com/smartroom/backend/repository/AuthenticationRepository.java
@@ -1,5 +1,6 @@
 package com.smartroom.backend.repository;
 
+import com.smartroom.backend.entity.Student;
 import com.smartroom.backend.entity.Teacher;
 
 import java.util.List;
@@ -15,4 +16,6 @@ public interface AuthenticationRepository {
     void deleteAllTeacher();
 
     void deleteAllStudent();
+
+    List<Student> fetchAllStudent();
 }
