+++ backend/src/main/java/com/smartroom/backend/repository/StudentRepository.java
@@ -1,8 +1,9 @@
 package com.smartroom.backend.repository;
 
+import com.smartroom.backend.entity.Student;
 import com.smartroom.backend.entity.StudentDetails;
 
 public interface StudentRepository {
 
-    StudentDetails getStudentById(String studentId);
+    Student getStudentById(String studentId);
 }
