+++ backend/src/main/java/com/smartroom/backend/service/StudentService.java
@@ -1,8 +1,9 @@
 package com.smartroom.backend.service;
 
+import com.smartroom.backend.entity.Student;
 import com.smartroom.backend.entity.StudentDetails;
 
 public interface StudentService {
 
-    StudentDetails getStudentDetails(String studentId);
+    Student getStudentDetails(String studentId);
 }
