+++ backend/src/main/java/com/smartroom/backend/service/TeacherServiceImpl.java
@@ -68,19 +68,11 @@ public class TeacherServiceImpl implements TeacherService {
     }
 
     @Override
-    public List<Student> fetchAllStudent() throws Exception {
-        try {
-            return teacherRepository.fetchAllStudent();
-        } catch (Exception e) {
-            throw new Exception(e.getLocalizedMessage());
-        }
-    }
-
-    @Override
     public String predictResult(String studentId, String subject) throws Exception {
         Student student = getStudentById(studentId);
-        if(student.getStudentDetails().getPredictedResult().get(subject) == null) {
+       if(student.getStudentDetails().getPredictedResult() == null || student.getStudentDetails().getPredictedResult().get(subject) == null) {
             HashMap<String, ArrayList<Integer>> studentMarks = student.getStudentDetails().getStudentMarks();
+            System.out.println("Student marks: "+studentMarks);
             if (!studentMarks.containsKey(subject) && !subject.equals("overall")) {
                 throw new InvalidParameter("Invalid subject,Subject not available");
             } else {
@@ -92,6 +84,7 @@ public class TeacherServiceImpl implements TeacherService {
                 }
 
                 try {
+                    System.out.println("Total marks: "+totalMarks);
                     LinkedHashMap<String, Object> params = getMLParams(student.getStudentDetails(), totalMarks);
                     String predictedResult =  getPrediction(params);
                     HashMap<String,String> storedPrediction = student.getStudentDetails().getPredictedResult();
@@ -175,13 +168,14 @@ public class TeacherServiceImpl implements TeacherService {
     private ArrayList<Integer> getTotalMarks(HashMap<String, ArrayList<Integer>> studentMarks) {
         ArrayList<Integer> totalMarks = new ArrayList<>();
         int term1 = 0, term2 = 0;
+        int totalSubject = studentMarks.size();
         for (Map.Entry entry : studentMarks.entrySet()) {
             ArrayList<Integer> marks = (ArrayList<Integer>) entry.getValue();
             term1 += marks.get(0);
             term2 += marks.get(1);
         }
-        totalMarks.add(term1);
-        totalMarks.add(term2);
+        totalMarks.add(term1/totalSubject);
+        totalMarks.add(term2/totalSubject);
 
         return totalMarks;
